Add changes to style.css as same as the activity
I was having a hard time figuring out how to
modify the title and h1 part, reason is when
removing all the padding and margin, it has its
own space allocated of ~10px

Additionally, when adding a margin in the elements
it only adds a space in left and right but the top
and bottom side of the card is missing. Adding a
space on those two result in bigger space than
expected.